Move logic in FormViewer outside return statement
